requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
redis==5.0.8
Jinja2==3.1.4
python-docx==1.1.2
email-validator==2.2.0  
bcrypt==4.2.0
python-multipart==0.0.12
pdfminer.six==20231228
reportlab==4.2.2
openai==1.51.0